Moved ExternalLink into replay-next package
