+++ packages/replay-next/components/ExternalLink.tsx
