+++ src/ui/components/Timeline/Capsule.tsx
@@ -2,9 +2,9 @@ import classNames from "classnames";
 import React, { useEffect, useRef, useState } from "react";
 import { CircularProgressbar, buildStyles } from "react-circular-progressbar";
 
+import ExternalLink from "replay-next/components/ExternalLink";
 import { useIndexingProgress } from "replay-next/src/hooks/useIndexingProgress";
 import { useGraphQLUserData } from "shared/user-data/GraphQL/useGraphQLUserData";
-import ExternalLink from "ui/components/shared/ExternalLink";
 import { useLoadingIsSlow } from "ui/components/Timeline/useLoadingIsSlow";
 import useModalDismissSignal from "ui/hooks/useModalDismissSignal";
 import { getBasicProcessingProgress, getShowFocusModeControls } from "ui/reducers/timeline";
