+++ src/ui/components/UploadScreen/Privacy.tsx
@@ -1,11 +1,11 @@
 import uniq from "lodash/uniq";
 import React, { Dispatch, SetStateAction } from "react";
 
+import ExternalLink from "replay-next/components/ExternalLink";
 import { OperationsData } from "shared/graphql/types";
 import { useGetRecording } from "ui/hooks/recordings";
 import { getRecordingId } from "ui/utils/recording";
 
-import ExternalLink from "../shared/ExternalLink";
 import MaterialIcon from "../shared/MaterialIcon";
 
 export function getUniqueDomains(operations: OperationsData) {
